Remove unneeded assertion and fix some simulation issues.
The assertion on #loadRelativeLiteralIn: had spurious triggers because the dependent may not be "inside the current compilation" when we are jit-ing a method at the end of the method zone. That trigger is spurious because then the method zone fails to allocate and nothing gets cogged.

The simulation errors were related to places where C says NULL is like zero (and smalltalk disagrees).